Move around add_cable_delay function to avoid circular import
